ssl/ech/ech_internal.c: avoid truncating time_t value in ossl_ech_status_print
Avoid casting time_t (that is almost universally 64-bit nowadays)
to (32-bit) int and cast it to long long instead.

Resolves: https://scan5.scan.coverity.com/#/project-view/65248/10222?selectedIssue=1681453
Fixes: ade10c5dbf20 "ECH client side"

+++ ssl/ech/ech_internal.c
@@ -804,8 +804,8 @@ void ossl_ech_status_print(BIO *out, SSL_CONNECTION *s, int selector)
                 OPENSSL_free(ec);
                 continue;
             }
-            BIO_printf(out, "ECH entry: %d public_name: %s age: %d%s\n",
-                i, pn, (int)secs, has_priv ? " (has private key)" : "");
+            BIO_printf(out, "ECH entry: %d public_name: %s age: %lld%s\n",
+                i, pn, (long long)secs, has_priv ? " (has private key)" : "");
             BIO_printf(out, "\t%s\n", ec);
             OPENSSL_free(pn);
             OPENSSL_free(ec);
